Move sources to a separate folder
