+++ xawscf/__init__.py
@@ -1,0 +1,5 @@
+'''
+Executes different commands.
+'''
+
+from .commands import commands as actions
