+++ xawscf/commands/deploy.py
@@ -4,11 +4,11 @@ Deploys cloudformation stack.
 
 import boto3
 import hashlib
-from utils import helper
-from utils.client import client
-from utils.cf_waiter import wait, WaiterError
-from utils.pattern import pattern
-from utils.logger import log
+from ..utils import helper
+from ..utils.client import client
+from ..utils.cf_waiter import wait, WaiterError
+from ..utils.pattern import pattern
+from ..utils.logger import log
 
 cf = client('cloudformation')
