+++ xawscf/commands/deploy_sources.py
@@ -3,11 +3,11 @@ Uploads zip archives to s3 bucket.
 '''
 
 from boto3.exceptions import S3UploadFailedError
-from utils import helper
-from utils.client import client
-from utils.pattern import pattern
-from utils.logger import log, logError
-from utils.parallel import run_parallel
+from ..utils import helper
+from ..utils.client import client
+from ..utils.pattern import pattern
+from ..utils.logger import log, logError
+from ..utils.parallel import run_parallel
 
 s3_client = client('s3')
