+++ xawscf/commands/info.py
@@ -2,9 +2,9 @@
 Gets cloudformation stack info.
 '''
 
-from utils.client import client, exceptions
-from utils.pattern import pattern
-from utils.logger import log
+from ..utils.client import client, exceptions
+from ..utils.pattern import pattern
+from ..utils.logger import log
 
 cf = client('cloudformation')
