+++ xawscf/commands/invoke.py
@@ -3,9 +3,9 @@ Invokes lambda function.
 '''
 
 import json
-from utils.client import client
-from utils.pattern import pattern
-from utils.logger import log, logError
+from ..utils.client import client
+from ..utils.pattern import pattern
+from ..utils.logger import log, logError
 
 lambda_client = client('lambda')
