+++ xawscf/commands/pack.py
@@ -6,10 +6,10 @@ import os
 from os import path
 import zipfile
 import shutil
-from utils import helper
-from utils.pattern import pattern
-from utils.logger import log
-from utils.parallel import run_parallel
+from ..utils import helper
+from ..utils.pattern import pattern
+from ..utils.logger import log
+from ..utils.parallel import run_parallel
 from .deploy_sources import run as call_deploy_sources
 
 def pack_directory(zf, real_dir, zip_dir):
