+++ xawscf/commands/process.py
@@ -2,10 +2,10 @@
 Translates pattern file into cloudformation template file.
 '''
 
-from utils import helper
-from utils.pattern import pattern
-from utils.logger import log
-from utils.yaml import save
+from ..utils import helper
+from ..utils.pattern import pattern
+from ..utils.logger import log
+from ..utils.yaml import save
 
 def run():
     log('Processing {}', helper.get_pattern_path())
