+++ xawscf/commands/statemachine.py
@@ -4,9 +4,9 @@ Starts (and cancels) state machine.
 
 import time
 import json
-from utils.client import session, client
-from utils.pattern import pattern
-from utils.logger import log, logError
+from ..utils.client import session, client
+from ..utils.pattern import pattern
+from ..utils.logger import log, logError
 
 stepfunctions = client('stepfunctions')
 sts = client('sts')
