+++ xawscf/pattern/base.py
