+++ xawscf/pattern/base_resource.py
