+++ xawscf/pattern/statemachine.py
@@ -1,5 +1,5 @@
 import json
-from utils.yaml import Custom
+from ..utils.yaml import Custom
 from .utils import get_full_name, make_output, try_set_field
 from .base_resource import BaseResource
 from .role import Role
