+++ xawscf/pattern/utils.py
